move exports out of page
