+++ util/sleep.tsx
@@ -1,0 +1,5 @@
+export default function sleep(ms = 2000) {
+  return new Promise((resolve) => {
+    setTimeout(resolve, ms);
+  });
+}
